requirements.txt
# basic requirements

protobuf>=4.25.3
transformers>=4.39.3
tokenizers>=0.15.0
cpm_kernels>=1.0.11
torch>=2.1.0
gradio>=4.26.0
sentencepiece>=0.2.0
sentence_transformers>=2.4.0
accelerate>=0.29.2
streamlit>=1.33.0
fastapi>=0.110.0
loguru~=0.7.2
mdtex2html>=1.3.0
latex2mathml>=3.77.0
jupyter_client>=8.6.1

# for openai demo
openai>=1.17.1
zhipuai>=2.0.1
pydantic>=2.7.0
sse-starlette>=2.0.0
uvicorn>=0.29.0
timm>=0.9.16
tiktoken>=0.6.0

# for langchain demo

langchain>=0.1.16
langchainhub>=0.1.15
arxiv>=2.1.0

# for web_ui
flask>=3.0.3
gevent>=24.2.1
flask_docs>=0.7.6
flask_cors>=4.0.0
flask_sqlalchemy>=3.1.1
simplejson>=3.19.2
pymysql>=1.1.0
PyMuPDF>=1.24.1
pymdown-extensions>=10.7.1

# for realted_work
evaluate>=0.4.0
jiwer>=3.0.3

# for kg
numpy==1.23.5
networkx==3.1
pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11
graphlearning==1.2.7
Annoy==1.17.2
pandas==2.0.3

matplotlib>=3.8.4
markdown>=3.6
pywebview>=5.1
python-dotenv>=1.0.1
werkzeug>=3.0.2
scipy>=1.13.0
scikit-learn>=1.4.2
requests>=2.31.0
lxml>=5.2.1